+++ lib/wax/attestation_statement_format.ex
@@ -13,6 +13,6 @@ defmodule Wax.AttestationStatementFormat do
   @callback verify(Wax.Attestation.statement(),
                    Wax.AuthenticatorData.t(),
                    Wax.ClientData.hash(),
-                   boolean())
+                   Wax.Challenge.t())
   :: {:ok, Wax.Attestation.result()} | {:error, any()}
 end