+++ lib/wax/attestation_statement_format/fido_u2f.ex
@@ -6,16 +6,16 @@ defmodule Wax.AttestationStatementFormat.FIDOU2F do
   @behaviour Wax.AttestationStatementFormat
 
   @impl Wax.AttestationStatementFormat
-  def verify(att_stmt, auth_data, client_data_hash, verify_trust_root) do
+  def verify(att_stmt, auth_data, client_data_hash, challenge) do
     with :ok <- valid_cbor?(att_stmt),
          {:ok, pub_key} <- extract_and_verify_public_key(att_stmt),
          public_key_u2f <- get_raw_cose_key(auth_data),
          verification_data <- get_verification_data(auth_data, client_data_hash, public_key_u2f),
          :ok <- valid_signature?(att_stmt["sig"], verification_data, pub_key),
-         :ok <-
-           (if verify_trust_root, do: attestation_certificate_valid?(att_stmt["x5c"]), else: :ok)
+         :ok <- attestation_certificate_valid?(att_stmt["x5c"], challenge)
     do
-      {attestation_type, metadata_statement} = determine_attestation_type(att_stmt["x5c"])
+      {attestation_type, metadata_statement} =
+        determine_attestation_type(att_stmt["x5c"], challenge)
 
       {:ok,
         {attestation_type,
@@ -97,13 +97,18 @@ defmodule Wax.AttestationStatementFormat.FIDOU2F do
     end
   end
 
-  @spec attestation_certificate_valid?([binary()]) :: :ok | {:error, any()}
+  @spec attestation_certificate_valid?([binary()], Wax.Challenge.t()) ::
+  :ok
+  | {:error, any()}
 
-  def attestation_certificate_valid?([leaf_cert | _]) do
+  def attestation_certificate_valid?(
+    [leaf_cert | _],
+    %Wax.Challenge{verify_trust_root: true} = challenge
+  ) do
     acki = Wax.Utils.Certificate.attestation_certificate_key_identifier(leaf_cert)
 
-    case Wax.Metadata.get_by_acki(acki) do
-      %Wax.MetadataStatement{} ->
+    case Wax.Metadata.get_by_acki(acki, challenge) do
+      %Wax.Metadata.Statement{} ->
         :ok
 
       nil ->
@@ -111,15 +116,20 @@ defmodule Wax.AttestationStatementFormat.FIDOU2F do
     end
   end
 
-  @spec determine_attestation_type([binary()])
-    :: {Wax.Attestation.type(), Wax.MetadataStatement.t() | nil}
+  def attestation_certificate_valid?(_, %Wax.Challenge{verify_trust_root: false}) do
+    :ok
+  end
 
-  defp determine_attestation_type([leaf_cert | _]) do
+  @spec determine_attestation_type([binary()], Wax.Challenge.t()) ::
+  {Wax.Attestation.type(), Wax.Metadata.Statement.t()}
+  | {Wax.Attestation.type(), nil}
+
+  defp determine_attestation_type([leaf_cert | _], challenge) do
     acki = Wax.Utils.Certificate.attestation_certificate_key_identifier(leaf_cert)
 
     Logger.debug("#{__MODULE__}: determining attestation type for acki=#{inspect(acki)}")
 
-    case Wax.Metadata.get_by_acki(acki) do
+    case Wax.Metadata.get_by_acki(acki, challenge) do
       nil ->
         {:uncertain, nil}