+++ lib/wax/attestation_statement_format/none.ex
@@ -4,7 +4,7 @@ defmodule Wax.AttestationStatementFormat.None do
   @behaviour Wax.AttestationStatementFormat
 
   @impl Wax.AttestationStatementFormat
-  def verify(_attstmt, _auth_data, _client_data_hash, _verify_trust_root) do
+  def verify(_attstmt, _auth_data, _client_data_hash, _challenge) do
     {:ok, {:none, nil, nil}}
   end
 end