+++ lib/wax/attestation_statement_format/packed.ex
@@ -28,23 +28,17 @@ defmodule Wax.AttestationStatementFormat.Packed do
   ]
 
   @impl Wax.AttestationStatementFormat
-  def verify(%{"x5c" => _} = att_stmt, auth_data, client_data_hash, verify_trust_root) do
+  def verify(%{"x5c" => _} = att_stmt, auth_data, client_data_hash, challenge) do
     with :ok <- valid_cbor?(att_stmt),
          :ok <- valid_x5c_signature?(att_stmt, auth_data, client_data_hash),
          :ok <- valid_attestation_certificate?(List.first(att_stmt["x5c"]), auth_data),
-         :ok <- (
-           if verify_trust_root do
-             attestation_path_valid?(att_stmt["x5c"], auth_data)
-           else
-             :ok
-           end
-         )
+         :ok <- attestation_path_valid?(att_stmt["x5c"], auth_data, challenge)
     do
       {:ok,
         {
-          determine_attestation_type(auth_data),
+          determine_attestation_type(auth_data, challenge),
           att_stmt["x5c"],
-          Wax.Metadata.get_by_aaguid(auth_data.attested_credential_data.aaguid)
+          Wax.Metadata.get_by_aaguid(auth_data.attested_credential_data.aaguid, challenge)
         }
       }
     else
@@ -190,14 +184,14 @@ defmodule Wax.AttestationStatementFormat.Packed do
     end
   end
 
-  @spec determine_attestation_type(Wax.AuthenticatorData.t()) :: Wax.Attestation.type()
+  @spec determine_attestation_type(Wax.AuthenticatorData.t(), Wax.Challenge.t()) :: Wax.Attestation.type()
 
-  defp determine_attestation_type(auth_data) do
+  defp determine_attestation_type(auth_data, challenge) do
     aaguid = auth_data.attested_credential_data.aaguid
 
     Logger.debug("#{__MODULE__}: determining attestation type for aaguid=#{inspect(aaguid)}")
 
-    case Wax.Metadata.get_by_aaguid(aaguid) do
+    case Wax.Metadata.get_by_aaguid(aaguid, challenge) do
       nil ->
         :uncertain
 
@@ -216,12 +210,17 @@ defmodule Wax.AttestationStatementFormat.Packed do
     end
   end
 
-  @spec attestation_path_valid?([binary()], Wax.AuthenticatorData.t())
-    :: :ok | {:error, any()}
+  @spec attestation_path_valid?([binary()], Wax.AuthenticatorData.t(), Wax.Challenge.t()) ::
+  :ok
+  | {:error, any()}
 
-  defp attestation_path_valid?(der_list, auth_data) do
-    case Wax.Metadata.get_by_aaguid(auth_data.attested_credential_data.aaguid) do
-      %Wax.MetadataStatement{attestation_root_certificates: arcs} ->
+  defp attestation_path_valid?(
+    der_list,
+    auth_data,
+    %Wax.Challenge{verify_trust_root: true} = challenge
+  ) do
+    case Wax.Metadata.get_by_aaguid(auth_data.attested_credential_data.aaguid, challenge) do
+      %Wax.Metadata.Statement{attestation_root_certificates: arcs} ->
         if Enum.any?(
           arcs,
           fn arc ->
@@ -245,5 +244,9 @@ defmodule Wax.AttestationStatementFormat.Packed do
       _ ->
         {:error, :attestation_packed_no_attestation_metadata_statement_found}
     end
+  end
+
+  defp attestation_path_valid?(_, _, %Wax.Challenge{verify_trust_root: false}) do
+    :ok
   end
 end